Fix hi-res floor/ceiling black bands (and likely crash): visplane rows were bytes
The visplane_t top[]/bottom[] arrays (per-column floor/ceiling clip rows) were
'byte', capping values at 255. At the base 320x200 that's fine (rows < 200),
but at higher internal resolutions view rows exceed 255, so floors/ceilings
could not be drawn below row 255 -- leaving black bands at the bottom of the
screen (most visible in fullscreen view, worse the higher the resolution, and
absent at 320x200, exactly as reported). The byte overflow also corrupted span
computation, the likely cause of the intermittent crash.

Widen top[]/bottom[] (and their pads) to unsigned short and use 0xffff as the
'unset' sentinel (was 0xff). The struct comment literally flagged this as 'the
rub for all dynamic resize/change of resolution'.

Verified floors render to the bottom in fullscreen view and normal view at
640x400 / 960x600 / 1280x800.

+++ r_defs.h
@@ -466,16 +466,16 @@ typedef struct
   int			maxx;
   
   // leave pads for [minx-1]/[maxx+1]
-  
-  byte		pad1;
-  // Here lies the rub for all
-  //  dynamic resize/change of resolution.
-  byte		top[MAXWIDTH];
-  byte		pad2;
-  byte		pad3;
-  // See above.
-  byte		bottom[MAXWIDTH];
-  byte		pad4;
+
+  // These hold screen row numbers, so they must be wide enough for the
+  // internal resolution (a byte only reaches 255 -> broke hi-res > 255 rows).
+  // 0xffff is the "unset" sentinel (see R_FindPlane / R_CheckPlane).
+  unsigned short	pad1;
+  unsigned short	top[MAXWIDTH];
+  unsigned short	pad2;
+  unsigned short	pad3;
+  unsigned short	bottom[MAXWIDTH];
+  unsigned short	pad4;
 
 } visplane_t;
